+++ go/api/server_management_commands_test.go
@@ -7,6 +7,7 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/google/uuid"
 	"github.com/valkey-io/valkey-glide/go/api/options"
 )
 
@@ -199,4 +200,29 @@ func ExampleGlideClient_LolwutWithOptions() {
 	// Output:
 	// LOLWUT version result is of type string
 	// LOLWUT version with args result is of type string
+}
+
+func ExampleGlideClient_LastSave() {
+	var client *GlideClient = getExampleGlideClient() // example helper function
+	key := "key-" + uuid.NewString()
+	client.Set(key, "hello")
+	response, err := client.LastSave()
+	if err != nil {
+		fmt.Println("Glide example failed with an error: ", err)
+	}
+	fmt.Println(response > 0)
+
+	// Output: true
+}
+
+func ExampleGlideClient_ConfigResetStat() {
+	var client *GlideClient = getExampleGlideClient() // example helper function
+	response, err := client.ConfigResetStat()
+	if err != nil {
+		fmt.Println("Glide example failed with an error: ", err)
+	}
+	fmt.Println(response)
+
+	// Output:
+	// OK
 }