+++ go/integTest/standalone_commands_test.go
@@ -814,6 +814,26 @@ func (suite *GlideTestSuite) TestLolwutWithOptions_EmptyArgs() {
 	assert.Contains(suite.T(), res, "Redis ver.")
 }
 
+func (suite *GlideTestSuite) TestClientId() {
+	client := suite.defaultClient()
+	result, err := client.ClientId()
+	assert.Nil(suite.T(), err)
+	assert.Greater(suite.T(), result, int64(0))
+}
+
+func (suite *GlideTestSuite) TestLastSave() {
+	client := suite.defaultClient()
+	t := suite.T()
+	result, err := client.LastSave()
+	assert.Nil(t, err)
+	assert.Greater(t, result, int64(0))
+}
+
+func (suite *GlideTestSuite) TestConfigResetStat() {
+	client := suite.defaultClient()
+	suite.verifyOK(client.ConfigResetStat())
+}
+
 func (suite *GlideTestSuite) TestRandomKey() {
 	client := suite.defaultClient()
 	// Test 1: Check if Echo command return the message